begin cleanup of code.
