+++ Mandatum.Generators/CommandHandlerGenerator.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Runtime.Serialization;
 using System.Text;
 using Mandatum.Generators.Objects;
-using Mandatum.Generators.SyntaxReceivers;
+using Mandatum.Generators.Syntax;
 using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CSharp;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
